[DebugInfo] PATCH 2/3: Duplicate logics regarding debug_value_addr
This patch replace all in-memory objects of DebugValueAddrInst with
DebugValueInst + op_deref, and duplicates logics that handles
DebugValueAddrInst with the latter. All related check in the tests
have been updated as well.

Note that this patch neither remove the DebugValueAddrInst class nor
remove `debug_value_addr` syntax in the test inputs.

+++ lib/SILOptimizer/Transforms/SILMem2Reg.cpp
@@ -88,7 +88,8 @@ static void replaceDestroy(DestroyAddrInst *dai, SILValue newValue,
 }
 
 /// Promote a DebugValueAddr to a DebugValue of the given value.
-static void promoteDebugValueAddr(DebugValueAddrInst *dvai, SILValue value,
+template <class DebugInstTy>
+static void promoteDebugValueAddr(DebugInstTy *dvai, SILValue value,
                                   SILBuilderContext &ctx,
                                   InstructionDeleter &deleter) {
   assert(dvai->getOperand()->getType().isLoadable(*dvai->getFunction()) &&
@@ -97,15 +98,27 @@ static void promoteDebugValueAddr(DebugValueAddrInst *dvai, SILValue value,
   // Avoid inserting the same debug_value twice.
   for (auto *use : value->getUses()) {
     if (auto *dvi = dyn_cast<DebugValueInst>(use->getUser())) {
-      if (*dvi->getVarInfo() == *dvai->getVarInfo()) {
+      // Since we're not comparing di-expression in
+      // SILDebugVariable::operator==(), it's necessary to distinguish
+      // debug_value w/ normal values from that with address-type values.
+      if (!dvi->hasAddrVal() &&
+          *dvi->getVarInfo() == *dvai->getVarInfo()) {
         deleter.forceDelete(dvai);
         return;
       }
     }
   }
 
+  auto VarInfo = *dvai->getVarInfo();
+  // Drop op_deref if dvai is actually a debug_value instruction
+  if (isa<DebugValueInst>(dvai)) {
+    auto &DIExpr = VarInfo.DIExpr;
+    if (DIExpr)
+      DIExpr.eraseElement(DIExpr.element_begin());
+  }
+
   SILBuilderWithScope b(dvai, ctx);
-  b.createDebugValue(dvai->getLoc(), value, *dvai->getVarInfo());
+  b.createDebugValue(dvai->getLoc(), value, std::move(VarInfo));
   deleter.forceDelete(dvai);
 }
 
@@ -426,6 +439,11 @@ StoreInst *StackAllocationPromoter::promoteAllocationInBlock(
         promoteDebugValueAddr(dvai, runningVal, ctx, deleter);
       continue;
     }
+    if (auto *dvi = DebugValueInst::hasAddrVal(inst)) {
+      if (dvi->getOperand() == asi && runningVal)
+        promoteDebugValueAddr(dvi, runningVal, ctx, deleter);
+      continue;
+    }
 
     // Replace destroys with a release of the value.
     if (auto *dai = dyn_cast<DestroyAddrInst>(inst)) {
@@ -627,6 +645,13 @@ void StackAllocationPromoter::fixBranchesAndUses(BlockSetVector &phiBlocks) {
       // Replace DebugValueAddr with DebugValue.
       SILValue def = getLiveInValue(phiBlocks, userBlock);
       promoteDebugValueAddr(dvai, def, ctx, deleter);
+      ++NumInstRemoved;
+      continue;
+    } else if (auto *dvi = DebugValueInst::hasAddrVal(user)) {
+      // Replace debug_value w/ address-type value with
+      // a new debug_value w/ promoted value.
+      SILValue def = getLiveInValue(phiBlocks, userBlock);
+      promoteDebugValueAddr(dvi, def, ctx, deleter);
       ++NumInstRemoved;
       continue;
     }
@@ -975,7 +1000,8 @@ static bool isCaptured(AllocStackInst *asi, bool &inSingleBlock) {
 
     // Deallocation is also okay, as are DebugValueAddr. We will turn
     // the latter into DebugValue.
-    if (isa<DeallocStackInst>(user) || isa<DebugValueAddrInst>(user))
+    if (isa<DeallocStackInst>(user) || isa<DebugValueAddrInst>(user) ||
+        DebugValueInst::hasAddrVal(user))
       continue;
 
     // Destroys of loadable types can be rewritten as releases, so
@@ -1011,7 +1037,8 @@ bool MemoryToRegisters::isWriteOnlyAllocation(AllocStackInst *asi) {
 
     // If we haven't already promoted the AllocStack, we may see
     // DebugValueAddr uses.
-    if (isa<DebugValueAddrInst>(user))
+    if (isa<DebugValueAddrInst>(user) ||
+        DebugValueInst::hasAddrVal(user))
       continue;
 
     if (isDeadAddrProjection(user))
@@ -1078,6 +1105,18 @@ void MemoryToRegisters::removeSingleBlockAllocation(AllocStackInst *asi) {
           assert(asi->getElementType().isVoid() &&
                  "Expected initialization of non-void type!");
           deleter.forceDelete(dvai);
+        }
+      }
+      continue;
+    } else if (auto *dvi = DebugValueInst::hasAddrVal(inst)) {
+      if (dvi->getOperand() == asi) {
+        if (runningVal) {
+          promoteDebugValueAddr(dvi, runningVal, ctx, deleter);
+        } else {
+          // Drop debug_value of uninitialized void values.
+          assert(asi->getElementType().isVoid() &&
+                 "Expected initialization of non-void type!");
+          deleter.forceDelete(dvi);
         }
       }
       continue;
